% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_analysis.r
\name{associated_gene_visualization}
\alias{associated_gene_visualization}
\title{Visulization of associated genes of tissue modules}
\usage{
associated_gene_visualization(res, quantile_thresh = 0.75)
}
\arguments{
\item{res}{the results of 'run_STModule' running on a tissue section.}

\item{quantile_thresh}{names will be demonstrated for genes with activity higher than the quantile threshold.}
}
\value{
a list of plots each for a tissue module
}
\description{
Visulization of associated genes of tissue modules
}
