% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cal_cov_inv_list}
\alias{cal_cov_inv_list}
\title{Calculate the inverse of covariance matrices for all tissue modules.}
\usage{
cal_cov_inv_list(r, dist_mat, params)
}
\arguments{
\item{r}{a list of variable r of the SE kernel for all tissue modules.}

\item{dist_mat}{Euclidean distance matrix of the spots/cells.}

\item{params}{parameters of the estimation process.}
}
\value{
a list of inverse matrices of the covariance matrices of all tissue moduels.
}
\description{
Calculate the inverse of covariance matrices for all tissue modules.
}
