% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cal_cov_inv_mat}
\alias{cal_cov_inv_mat}
\title{Calculate inverse of covariance matrix}
\usage{
cal_cov_inv_mat(r, dist_mat, params)
}
\arguments{
\item{r}{variable r of the SE kernel.}

\item{dist_mat}{Euclidean distance matrix of the spots/cells.}

\item{params}{parameters of the estimation process.}
}
\value{
inverse matrix of the covariance matrix.
}
\description{
Calculate inverse of covariance matrix
}
