% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cal_distance_mat}
\alias{cal_distance_mat}
\title{Calculate Euclidean distance matrix of spots/cells based on location information.}
\usage{
cal_distance_mat(locations, dimension = "2D")
}
\arguments{
\item{locations}{location information of spots/cells.}

\item{dimension}{dimension of the coordinates. '2D' or '3D'.}
}
\value{
a Euclidean distance matrix of spots/cells.
}
\description{
Calculate Euclidean distance matrix of spots/cells based on location information.
}
