% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cal_mat_inv}
\alias{cal_mat_inv}
\title{Calculate matrix inverse}
\usage{
cal_mat_inv(mat, params)
}
\arguments{
\item{mat}{matrix to calculate inverse.}

\item{params}{parameters of the estimation process.}
}
\value{
inverse matrix of the input matrix.
}
\description{
Calculate matrix inverse
}
