% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cal_precision_log_det}
\alias{cal_precision_log_det}
\title{Calculate logarithm of determinant of a precision matrix.}
\usage{
cal_precision_log_det(prec_mat)
}
\arguments{
\item{prec_mat}{a precision matrix for spatial maps.}
}
\value{
the logarithm of determinant of a precision matrix.
}
\description{
Calculate logarithm of determinant of a precision matrix.
}
