% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{data_filtering}
\alias{data_filtering}
\title{Filtering input data, removing genes and spots/cells by the criteria specified by the parameters}
\usage{
data_filtering(count_mat, high_resolution, gene_filtering, cell_thresh = 100)
}
\arguments{
\item{count_mat}{count matrix of input data}

\item{high_resolution}{a boolean value to indicate spatial resolution of the data.
\itemize{
\item TRUE for data profiled by 'Slide-seq', 'Slide-seqV2', 'Stereo-seq', etc.
\item FALSE for data profiled by 'ST' and '10x Visium'.
}}

\item{gene_filtering}{parameter to filter the genes, removing genes expressed in less than gene_filtering locations/cells.
\itemize{
\item when 'high_resolution' is FALSE: gene_filtering indicates the threshold of percentage of locations.
\item when 'high_resolution' is TRUE: gene_filtering indicates the threshold of number of cells.
}}

\item{cell_thresh}{parameter to filter cells for high-resolution data, removing cells with less than cell_thresh counts.}
}
\value{
the count matrix after filtering
}
\description{
Filtering input data, removing genes and spots/cells by the criteria specified by the parameters
}
