% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.r
\name{data_preprocessing}
\alias{data_preprocessing}
\title{Data preprocessing, to generate the input data for STModule from the user-provided data.
input data: count matrix & spatial information
\itemize{
\item count matrix: locations * genes. rownames - id of spots/cells; colnames - gene names.
\item spatial information: locations * coordinates. rownames - id of spots/cells; colnames - x, y.
}}
\usage{
data_preprocessing(
  count_file,
  loc_file,
  high_resolution = FALSE,
  gene_mode = "HVG",
  top_hvg = 2000,
  gene_list = c(),
  file_sep = "\\t",
  gene_filtering = 0.1,
  cell_thresh = 100
)
}
\arguments{
\item{count_file}{path and file name of the count matrix.}

\item{loc_file}{path and file name of the spatial information.}

\item{high_resolution}{a boolean value to indicate spatial resolution of the data.
\itemize{
\item TRUE for data profiled by 'Slide-seq', 'Slide-seqV2', 'Stereo-seq', etc.
\item FALSE for data profiled by 'ST' and '10x Visium'.
}}

\item{gene_mode}{the way to select genes for the analysis.
\itemize{
\item 'HVG': to use top highly variable genes selected by Seurat.
\item 'selected': to use user-selected genes included in the filtered data, provided by the 'gene_list' parameter.
\item 'combined': to use the combination of top HVGs and user-selected genes.
}}

\item{top_hvg}{the number of top HVGs to use in the analysis, default 2000.}

\item{gene_list}{a vector of genes provided by the user to use in the analysis.}

\item{file_sep}{the field separator character.}

\item{gene_filtering}{parameter to filter the genes, removing genes expressed in less than gene_filtering locations/cells.
\itemize{
\item when 'high_resolution' is FALSE: gene_filtering indicates the threshold of percentage of locations.
\item when 'high_resolution' is TRUE: gene_filtering indicates the threshold of number of cells.
}}

\item{cell_thresh}{parameter to filter cells for high-resolution data, removing cells with less than cell_thresh counts.}
}
\value{
a list containing the processed expression matrix, distance matrix and location information.
}
\description{
Data preprocessing, to generate the input data for STModule from the user-provided data.
input data: count matrix & spatial information
\itemize{
\item count matrix: locations * genes. rownames - id of spots/cells; colnames - gene names.
\item spatial information: locations * coordinates. rownames - id of spots/cells; colnames - x, y.
}
}
