% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_analysis.r
\name{get_assocaited_genes}
\alias{get_assocaited_genes}
\title{Get associated genes of tissue modules}
\usage{
get_assocaited_genes(res)
}
\arguments{
\item{res}{the results of 'run_STModule' running on a tissue section.}
}
\value{
a dataframe of member information of the tissue modules. Columns: 'gene', 'activity', 'module'.
}
\description{
Get associated genes of tissue modules
}
