% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_prec_mat_diag}
\alias{get_prec_mat_diag}
\title{Get diagonal of precision matrices.}
\usage{
get_prec_mat_diag(prec_list, dimension, num_comp)
}
\arguments{
\item{prec_list}{a list of precision matrices of spatial maps.}

\item{dimension}{dimension of precision matrix.}

\item{num_comp}{number of tissue modules.}
}
\value{
diagonal of precision matrices.
}
\description{
Get diagonal of precision matrices.
}
