% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{remove_zero_exp_locations}
\alias{remove_zero_exp_locations}
\title{Removing spots/cells with zero expression}
\usage{
remove_zero_exp_locations(count_mat)
}
\arguments{
\item{count_mat}{count matrix}
}
\value{
the count matrix after filtering
}
\description{
Removing spots/cells with zero expression
}
