% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_STModule.r
\name{run_spatial_map_estimation}
\alias{run_spatial_map_estimation}
\title{Estimate spatial maps of tissue module for a new tissue section}
\usage{
run_spatial_map_estimation(
  res,
  count_file,
  loc_file,
  high_resolution = FALSE,
  file_sep = "\\t",
  cell_thresh = 100,
  max_iter = 2000,
  version = "cpu"
)
}
\arguments{
\item{res}{the result of function 'run_STModule' on a tissue section}

\item{count_file}{path and file name of the count matrix of the new data}

\item{loc_file}{path and file name of the spatial information of the new data}

\item{high_resolution}{a boolean value to indicate spatial resolution of the data.
\itemize{
\item TRUE for data profiled by 'Slide-seq', 'Slide-seqV2', 'Stereo-seq', etc.
\item FALSE for data profiled by 'ST' and '10x Visium'.
}}

\item{file_sep}{the field separator character of the files for the new data.}

\item{cell_thresh}{parameter to filter cells for high-resolution data, removing cells with less than cell_thresh counts (default 100).}

\item{max_iter}{number of maximum iteration.}

\item{version}{to use CPU or GPU version of STModule.
\itemize{
\item 'cpu' for ST data.
\item 'gpu' for 10x Visium and high-resolution data.
}}
}
\value{
a list containing estimated results for variables in the model, as well as parameters, gene list, location list and location information.
}
\description{
Estimate spatial maps of tissue module for a new tissue section
}
