% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_STModule.r
\name{run_STModule}
\alias{run_STModule}
\title{Run STModule on spatially resolved transcriptomics data}
\usage{
run_STModule(
  data,
  num_modules,
  high_resolution = FALSE,
  max_iter = 2000,
  version = "cpu"
)
}
\arguments{
\item{data}{generated data after data preprocessing using the function data_preprocessing().}

\item{num_modules}{number of tissue modules to identify.}

\item{high_resolution}{a boolean value to indicate spatial resolution of the data.
\itemize{
\item TRUE for data profiled by 'Slide-seq', 'Slide-seqV2', 'Stereo-seq', etc.
\item FALSE for data profiled by 'ST' and '10x Visium'.
}}

\item{max_iter}{number of maximum iteration.}

\item{version}{to use CPU or GPU version of STModule.
\itemize{
\item 'cpu' for ST data.
\item 'gpu' for 10x Visium and high-resolution data.
}}
}
\value{
a list containing estimated results for variables in the model, as well as parameters, gene list, location list and location information.
}
\description{
Run STModule on spatially resolved transcriptomics data
}
