% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_analysis.r
\name{spatial_expression_visualization}
\alias{spatial_expression_visualization}
\title{Plot spatial expression of interested genes}
\usage{
spatial_expression_visualization(
  count_file,
  loc_file,
  gene_list,
  output_dir,
  file_sep = "\\t",
  point_size = 6
)
}
\arguments{
\item{count_file}{path and file name of the count matrix}

\item{loc_file}{path and file name of the spatial information}

\item{gene_list}{genes to visualize}

\item{output_dir}{output directory of the plots}

\item{file_sep}{the field separator character.}

\item{point_size}{size of points in the spatial expression plot
\itemize{
\item ST data: point_size = 6
\item 10x Visium data: point_size = 2
\item Slide-seqV2 and Stereo-seq data: point_size = 0.3
}}
}
\value{
pdf files in the specified output directory, each for the spatial expression of a query gene
}
\description{
Plot spatial expression of interested genes
}
