% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_factorization_gpu.r
\name{spatial_factorization_gpu}
\alias{spatial_factorization_gpu}
\title{Identify tissue modules, GPU version}
\usage{
spatial_factorization_gpu(
  params,
  profile,
  dist_mat,
  maxiter = 2000,
  track = 10,
  debugging = "no",
  gpumatirx_type = "torch",
  pip_thresh = 1
)
}
\arguments{
\item{params}{parameters of the estimation process.}

\item{profile}{pre-processed expression matrix.}

\item{dist_mat}{Euclidean distance matrix of spots/cells calculated from location information.}

\item{maxiter}{number of maximum iteration.}

\item{track}{number of iterations to track the changes of tissue module members.}

\item{debugging}{the way to output FE information for debugging
\itemize{
\item 'no': not output FE information for debugging
\item 'each_update': output FE information in each iteration
\item '10_iter': output FE information for each 10 iteration
}}

\item{gpumatirx_type}{type of GPU matrix.
\itemize{
\item 'torch': using torch tensors.
\item 'tensorflow': using tensorflow tensors.
}}

\item{pip_thresh}{parameter for stopping criteria. Number of changes of tissue module members.}
}
\value{
a list of estimated results for the variables in the model.
}
\description{
Identify tissue modules, GPU version
}
