% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_map_estimation.r
\name{spatial_map_estimation}
\alias{spatial_map_estimation}
\title{Estimating spatial maps of tissue modules on another tissue section, CPU version}
\usage{
spatial_map_estimation(
  params,
  profile,
  dist_mat,
  module_loadings,
  maxiter = 2000,
  track = 10,
  debugging = "iter_count",
  iter_count = 1,
  pip_thresh = 1
)
}
\arguments{
\item{params}{parameters of the estimation process.}

\item{profile}{pre-processed expression matrix.}

\item{dist_mat}{Euclidean distance matrix of spots/cells calculated from location information.}

\item{module_loadings}{members of tissue modules.}

\item{maxiter}{number of maximum iteration.}

\item{track}{number of iterations to track the changes of tissue module members.}

\item{debugging}{the way to output FE information for debugging
\itemize{
\item 'no': not output FE information for debugging
\item 'each_update': output FE information in each iteration
\item 'iter_count': output FE information for each 'iter_count' iteration
}}

\item{iter_count}{number of iterations to check FE for debugging mode of 'iter_count'.}

\item{pip_thresh}{parameter for stopping criteria. Number of changes of tissue module members.}
}
\value{
a list of estimated results for the variables in the model.
}
\description{
Estimating spatial maps of tissue modules on another tissue section, CPU version
}
