% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_analysis.r
\name{spatial_map_visualization}
\alias{spatial_map_visualization}
\title{Visiualization of spatial maps}
\usage{
spatial_map_visualization(res, normalization = "log", point_size = 6)
}
\arguments{
\item{res}{the results of 'run_STModule' or 'run_spatial_map_estimation' running on a tissue section.}

\item{normalization}{normalization method of spatial maps.
\itemize{
\item 'log': log-transformation by keeping the direction of activities, default.
\item 'scale': scale the spatial map of each module to a vector with zero-mean and unit-variance.
}}

\item{point_size}{size of points in the spatial map plot
\itemize{
\item ST data: point_size = 6
\item 10x Visium data: point_size = 2
\item Slide-seqV2 and Stereo-seq data: point_size = 0.3
}}
}
\value{
a list of plots each for a tissue module
}
\description{
Visiualization of spatial maps
}
